Docs & Example for "Roll Your Own JS Runtime" blog posts (#824)
I was surprised to find that one of the blog posts [provided docs][1]
for the `CreateSnapshotOptions` instead of just linking to its Rustdoc.
That's because the type itself [didn't have docs][2].

The first commit here is me just inlining those docs into Rustdocs.

But then, as I tried following the example, I found that the APIs have
diverged significantly from the blog post series. So I've added another
commit which adds an example project that can serve as a live/working
example of creating a snapshot for JsRuntime. Hopefully, since this will
get built with `cargo build`, it'll stay more up-to-date than the blog
posts.

If this goes too far, I'm happy to create a separate PR that is just the
first commit.

If you'd like me to go further, I think it might be a good idea to split
the extension out into a separate `ext` crate, which I think allows for
code re-use between `build.rs` and `main.rs`.

[1]:
https://deno.com/blog/roll-your-own-javascript-runtime-pt3#diving-into-createsnapshotoptions
[2]:
https://docs.rs/deno_core/0.294.0/deno_core/snapshot/struct.CreateSnapshotOptions.html


commit eb35c83775fe98f90816264ce6fee2632a03fc37

    Add docs for create_snapshot() and its options.

commit 618d423cbeedadd528930ec59221101e80533e05

    Add an example for "Roll Your Own JS Runtime"
    
    The blog post series had become out-of-date.
    
    Committing example code will hopefully bit-rot less?

---------

Co-authored-by: snek <the@snek.dev>
Co-authored-by: Ian Bull <irbull@gmail.com>